requirements.txt
requests
beautifulsoup4
scikit-learn
numpy
pandas
python-dotenv
